Finish pembuatan switch budget

+++ database/migrations/2021_12_14_201135_create_budget_table.php
@@ -17,7 +17,7 @@ class CreateBudgetTable extends Migration
             $table->id();
             $table->unsignedBigInteger('jenis_belanja_id');
             $table->year('tahun_anggaran');
-            $table->double('nominal');
+            $table->decimal('nominal', 18, 0);
             $table->text('keterangan')->nullable()->default(null);
             $table->timestamps();
